Moved Bio/Links to Profile-Info component
