+++ client/src/app/pages/dashboard/dashboard.module.ts
@@ -18,7 +18,7 @@ import {SmartTableModule} from './smart-table/smart-table.module';
 import { UserEditModalComponent } from './user-list/user-edit-modal/user-edit-modal.component';
 import { VisitorsStatisticsComponent } from './follow-runs/visitors-statistics/visitors-statistics.component';
 import { ProgressSectionComponent } from './follow-maps/progress-section.component';
-import {UserProfileModule} from './user-profile/user-profile.module';
+import {UserProfileModule} from './profile/user-profile.module';
 
 @NgModule({
   imports: [
