+++ client/src/app/pages/dashboard/profile-info/profile-info.component.html
@@ -1,0 +1,19 @@
+<div class="col-sm-12">
+  <nb-card>
+        <nb-card-body>
+            <h4>Bio:</h4>
+            <!--  <p>{{ user?.getBio }}</p> -->
+            <p>I've been a fan of the game since its development. I love making different kinds of maps so be sure to check them out!</p>
+
+          <h4>Links:</h4>
+          <div class="socials">
+            <a href="https://twitter.com/momentummod" target="_blank" class="ion ion-social-twitter"></a>
+            <a href="https://discordapp.com/invite/wQWkRb6" target="_blank">
+              <img src="assets\images\Discord-Logo-White.png" height="35px" width="35px"></a>
+          </div>
+        </nb-card-body>
+  </nb-card>
+</div>
+
+
+
