+++ client/src/app/pages/dashboard/profile/user-profile.component.html
@@ -1,0 +1,43 @@
+<div class= "row">
+  <div class="col-sm-6">
+    <ngx-profile-card></ngx-profile-card>
+  </div>
+  <div class="col-sm-6">
+    <profile-info></profile-info>
+  </div>
+</div>
+
+  <div class = "row">
+
+    <div class="col-sm-2">
+      <nb-card>
+        <nb-card-header>
+          <h4>Player Stats</h4>
+        </nb-card-header>
+        <nb-card-body>
+          <h6>Maps Completed:</h6> <p>37</p>
+          <h6>Totals Jumps:</h6> <p>411</p>
+          <h6>Total Strafes:</h6> <p>102</p>
+        </nb-card-body>
+      </nb-card>
+    </div>
+
+    <div class="col-sm-10">
+
+      <nb-card size="medium">
+        <nb-card-header>
+          <h5>Recent Maps</h5>
+        </nb-card-header>
+        <nb-card-body>
+        </nb-card-body>
+      </nb-card>
+      <nb-card size="medium">
+        <nb-card-header>
+          <h5>Recent Runs</h5>
+        </nb-card-header>
+        <nb-card-body>
+        </nb-card-body>
+      </nb-card>
+
+    </div>
+  </div>
