+++ client/src/app/pages/dashboard/profile/user-profile.module.ts
@@ -3,6 +3,7 @@ import { UserProfileComponent } from './user-profile.component';
 import { NgModule } from '@angular/core';
 import { ThemeModule } from '../../../@theme/theme.module';
 import { ProfileCardComponent } from '../profile-card/profile-card.component';
+ import { ProfileInfoComponent } from '../profile-info/profile-info.component';
 
 
 @NgModule({
@@ -12,10 +13,12 @@ import { ProfileCardComponent } from '../profile-card/profile-card.component';
   declarations: [
     ProfileCardComponent,
     UserProfileComponent,
+    ProfileInfoComponent,
   ],
   exports: [
     ProfileCardComponent,
     UserProfileComponent,
+    ProfileInfoComponent,
   ],
   providers: [],
 })
